I really need to make sure my commits are small.
There are a lot of code style changes.
More work on converting everything to a master dictionary. Some critical things had to be commented out because they rely on lookup methods that no longer exist. This is very much a work in progress, and commits will be smaller and typically incomplete.
Set the target framework to no longer build only for Windows and fixed indentation in the project file.
Established a convention for naming things that reference item categories to call them category indices.